ci: update & rename `build.yaml` workflow to `release.yaml`
This lays the groundwork for a release workflow using Semantic Release.
Adds a precursor job to both platform-specific build jobs that determine the release version and set it as a GitHub environment variable.
This uses a dry-run of semantic-release with the export plugin.

Adds placeholder steps to both workflows to apply this version before building.

Updates `actions/checkout` to `@v3` in both build jobs.

Adds a new `release` job, dependent on both builds, that publishes a new GitHub release with the built artifacts.
Contains placeholder to increment version (so that it can be committed with the new version tag)
Calls semantic release to achieve this, config as yet unset.